[Inspection Service] Remove the Debug Service!
