+++ Cargo.toml
@@ -44,7 +44,6 @@ members = [
     "crates/aptos-keygen",
     "crates/aptos-log-derive",
     "crates/aptos-logger",
-    "crates/aptos-metrics",
     "crates/aptos-metrics-core",
     "crates/aptos-proptest-helpers",
     "crates/aptos-rate-limiter",
@@ -59,8 +58,8 @@ members = [
     "crates/bounded-executor",
     "crates/channel",
     "crates/crash-handler",
-    "crates/debug-interface",
     "crates/fallible",
+    "crates/inspection-service",
     "crates/num-variants",
     "crates/proxy",
     "crates/short-hex-str",
