--- crates/debug-interface/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "debug-interface"
-version = "0.1.0"
-authors = ["Aptos Labs <opensource@aptoslabs.com>"]
-description = "Aptos debug interface"
-repository = "https://github.com/aptos-labs/aptos-core"
-homepage = "https://aptoslabs.com"
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-anyhow = "1.0.57"
-bytes = "1.1.0"
-reqwest = { version = "0.11.10", features = ["blocking", "json"], default_features = false }
-serde = { version = "1.0.137", features = ["derive"], default-features = false }
-tokio = { version = "1.18.2", features = ["full"] }
-warp = "0.3.2"
-
-aptos-config = { path = "../../config" }
-aptos-logger = { path = "../../crates/aptos-logger" }
-aptos-metrics = { path = "../../crates/aptos-metrics" }
-aptos-workspace-hack = { path = "../aptos-workspace-hack" }
