+++ crates/inspection-service/Cargo.toml
@@ -1,24 +1,27 @@
 [package]
-name = "aptos-metrics"
+name = "inspection-service"
 version = "0.1.0"
 authors = ["Aptos Labs <opensource@aptoslabs.com>"]
-description = "Aptos metrics utilities"
+description = "The Node Inspection Service"
 repository = "https://github.com/aptos-labs/aptos-core"
 homepage = "https://aptoslabs.com"
 license = "Apache-2.0"
 publish = false
 edition = "2018"
-build = "build.rs"
 
 [dependencies]
+anyhow = "1.0.57"
 futures = "0.3.21"
 hyper = { version = "0.14.18", features = ["full"] }
 once_cell = "1.10.0"
 prometheus = { version = "0.13.0", default-features = false }
+reqwest = { version = "0.11.10", features = ["blocking", "json"], default_features = false }
 serde_json = "1.0.81"
 sysinfo = "0.24.2"
 tokio = { version = "1.18.2", features = ["full"] }
 
+aptos-config = { path = "../../config" }
+aptos-infallible = { path = "../../crates/aptos-infallible" }
 aptos-logger = { path = "../../crates/aptos-logger" }
 aptos-metrics-core = { path = "../aptos-metrics-core" }
 aptos-telemetry = { path = "../aptos-telemetry" }
