+++ crates/inspection-service/src/json_encoder.rs
